4. Testes, segurança e lapidações/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
Django==4.1.4
django-admin-honeypot==1.1.0
django-cors-headers==3.13.0
django-redis==5.2.0
djangorestframework==3.14.0
Pillow==9.3.0
pytz==2022.6
redis==4.4.0
sqlparse==0.4.3
tzdata==2022.7
validate-docbr==1.10.0
